tests pass for checkers
