+++ tests/data/schema.json
@@ -2,6 +2,9 @@
   "type": "object",
   "properties": {
     "isPositive": { "type": "number" },
+    "isNegative": { "type": "number" },
+    "always": { "type": "string" },
+    "flagged": { "type": "number" },
     "additionalProperties": false
   }
 }
